Add AND/OR, fix parsing bugs, fix NOT return value
LOTS of work went into getting everything to work!
For a long time, I had the choice of:

1)
5>4>3 works, but (5>4>3) == 1 is treated as 5>4>3==1 which is False.

2)
5>4>3, (5>4>3)==1 and 1==(5>4>3) all work, but 1 == 2 || 3 == 4 yields True.
It was parsed as 1 == (2 || 3) == 1, despite the precedence being set correctly.
(The grammar rules were the issue.)

After plenty of work, I believe the solution was the rewrite in comparison chaining.
Previously, the rule was "comp : exp compop exp", which breaks in "1 == 2 || 3 == 1",
as the parser reduces the expressions in the wrong order.
(compop is == != > < >= <=)

The new rules are instead "comp : exp compop exp" and "comp : comp compop exp",
where "exp : comp". I assume the reason this works is because it finds the second rule
*before* reducing the comp to exp.

With that working, I could remove an ugly hack to make the (5>4>3) == 1 case work; that hack
seems to be what caused the issues with OR and AND.