Add functions, return statements and scopes!!
With this commit, user-defined functions (including nested functions) work!
Global variables (and variables in parent scopes) can be read and modified;
if a write occurs to a variable that does not exist in ANY scope, a local one
is created.

Thus, the following code works exactly as intended:

loops = 0;
func fac(n) {
	loops = loops + 1;
	func test(x) { print("Nested function says:", x); }
    if n < 2 { return 1; }
    else { test(n); return n * fac(n - 1); }
}

print("fac(5) =", fac(5));
print("number of loops:", loops);