Rename project to Procyon
The previous name (excalc) was chosen when the goal was to create a program
that could evaluate mathematical expressions, like (1+2)*3^4, with the correct
order of operations. I then added support for mathematical functions and comparisons,
and eventually user-defined functions, if statements and so on...
It's not exactly a calculator any more.

Enter the Procyon programming language!