Implement abort() and continue; rework control flow exceptions
After already having ReturnException and BreakException, I felt that adding
both ContinueException and AbortException would be a bit silly (since intr/repl
would have to have FOUR separate except blocks for four so similar things), so
I reworked them into ProcyonControlFlowException, that has a dictionary as its
argument.
"type" in the dictionary is either "return", "break", "continue" or "abort";
if it is "return", the dictionary also has a "value" key.