Merge repl.py and intr.py into procyon.py.
When called without arguments, the REPL is started.
When called with exactly one argument, it is treated as a file to interpret.
Any other number of arguments is invalid, and causes a usage message to be printed.

Most likely, some form of argv support will be added in the future... but until then,
multiple arguments make little sense.

Closes #18.