Move _ from the interpreter to the REPL; update readme.
Previously, _ was set to the last computed value, for every single evaluation in the program.
This causes a LOT of assignments, the VAST majority of which will go unused. This caused a
10-15% performance hit in a simple test, so I decided to move that logic to the interpreter instead,
as that is the only place where it is likely to be used, anyway!

Now, the interpreter sends along the last result from the previous evaluation to the interpreter,
which assigns the _ variable to that value for the entire next evaluation.