Add else if-blocks!
While the parsing code looks fairly long, most of it is actually comments and docstrings.
6 lines of actual code is removed from parser.py, and 21 new added, so the net is really only +15 lines.

There is also new only a single parser rule for if statementments (though it uses two helper rules for else if),
instead of the previous two: one for if {}, one for if {} else {}.

Note that no change to the interpreter was required! The lines that are changed there in
this commit merely changes the return value, to avoid results like [[[None]]] being returned to the REPL
during deeply nested if statements; the logic is unchanged.
The previous lines were in these in case I made if statements actually return something (as in e.g. Rust),
but I believe I will leave them as statements.

Closes #2.