Rewrite scoping rules and add global variables ($name); misc. clean-up
Most notably, this commit changes the scoping rules significantly.
Previously, assigning to a non-local variable would, if it existed in a parent scope (including
the global scope), assign to that pre-existing variable.
This had unintended consequences such as using "i" as a looping variable in a function
overwriting a different "i" used in the global scope.

The new rules (and the old ones) are verbosely documented in interpreter.py.

In short:

* Reading works as previously; if a function doesn't ever assign to a variable, but it
exists in an outer scope, the one in the outer scope is read.

* Assignment does not at all work as before. Assignment is ALWAYS to the local scope, possibly
shadowing variables in outer scopes.

* Global variables, which begin with a $, ignore all of these rules, and are simply stored in the
global scope no matter what. As such, creating such a variable inside a function and later
accessing it from another function or in the global scope is perfectly OK -- but the usual
caveats with global variables apply, of course, so they shouldn't be used very often.

The misc. clean-up includes deleting repl.py (I thought it was gone long ago, but apparenently I
forgot to delete it during the intr.pl/repl.pr -> procyon.py merger) and some whitespace/comment
edits here and there.

Closes #19.