Fix syntax error pointer when lines contain tabs.
A pointer is printed at syntax errors, like so:
	a = 12 + 3!;
              ^
Syntax error: unexpected NOT at ...

Previously, a fixed number of spaces were printed prior to the caret, namely
the number of characters that precededed the error token.
Now, the same whitespace that was used on the line itself is used.

This also takes care of issue #12, as I have now make the appropriate tests to ensure
that comments do not screw this up, either.

Also, fix syntax error capitalization to be in line with other exceptions;
the messages are printed in the middle of a sentence, and so should not be capitalized.

Closes #7.
Closes #12.