Allow multi-line input in the REPL, if the current line cannot be interpreted as-is.
For example, ending a line with ( or + that would cause an "unexpected end of input" error
now causes the REPL to keep reading further lines.

Since the language is evaluated, it makes no difference to evaluate line a first, immediately followed by line b, or to evaluate lines a+b at the "same time" (from the user's perspective).
Therefore, if a line *can* be evaluated, i.e. it is proper syntax, it *will* be.

What this means in practice is that you should not end a line with enter if you don't want it
to run; if you want to continue, end with e.g. an operator, an open paren and so on.

This commit also changes ^C to simply abort the current input, instead of exiting entirely.
As before, ^D (end-of-file) exits the interpreter. Same as with many other languages, such as
Python and Ruby.

Closes #4.