Tab-complete filenames in .import statements.
Note that tab completion is ONLY active when line begins with ".import".
I attempted to hack the completer such that it adds a tab in all other cases,
so that tab can be used for indentation when "coding" in the REPL,
yet still be used for tab completing filenames when importing.

Closes #16.