Use the new AST design to print better error messages.
Now that positional information is available for all tokens, we can use that
to print *where* errors of (almost) all types appear, rather than just for
syntax errors, as it was previously.

For example, having the addition a = "hello"; ... 1 + a: ... in the middle of
a long file will now print the offending line and even operator, rather than
simply printing a very nonspecific invalid addition line.

Errors without locational info are overflow errors and procyon internal errors
(which have no location in the program itself, anyway, but one in the interpreter).

Closes #23.